refactor: growl and title
