+++ packages/Toast/Growl.js
@@ -3,7 +3,6 @@ import { func, node } from 'prop-types'
 import { ClearButton } from '@welcome-ui/clear-button'
 
 import * as S from './styles'
-import { Title } from './Title'
 
 export const Growl = forwardRef(({ children, dataTestId, onClose }, ref) => (
   <S.Growl ref={ref}>
@@ -27,6 +26,3 @@ Growl.propTypes /* remove-proptypes */ = {
   children: node.isRequired,
   onClose: func
 }
-
-Growl.Action = S.Action
-Growl.Title = Title
