--- packages/Toast/Snackbar/styled.js
@@ -1,11 +1,0 @@
-import styled from '@xstyled/styled-components'
-import { Alert } from '@welcome-ui/alert'
-
-export const Snackbar = styled(Alert)`
-  display: flex;
-  align-items: center;
-
-  & > *:not(:only-child):not(:last-child) {
-    margin-right: sm;
-  }
-`
