+++ packages/Toast/Title.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 
-import { render } from '../../../src/utils/tests'
-import { colors } from '../../Core/theme/colors'
+import { render } from '../../src/utils/tests'
+import { colors } from '../Core/theme/colors'
 
 import { Title } from './Title'
